#26 Add view for showing stored emergency contacts
