--- android/app/src/main/java/platis/solutions/smartcarmqttcontroller/EmergencyContacts.java
@@ -1,14 +1,0 @@
-package platis.solutions.smartcarmqttcontroller;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class EmergencyContacts extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_emergency_contacts);
-    }
-}
